api: major fix to Get(), Min(), Max() APIs.
Instead of return a Node object these APIs are not accepting
a node-callback. The node object is valid only till callback
returns and this can help in reducing garbage.

These APIs now return a boolean value, where if,
true  - key is present and returns the callback's return value
true  - key is present and callback is nil
false - key is missing