api: UpsertCAS api added.
UpsertCAS will take a CAS argument. Every document shall have a seqno that
always increases, not necessariy monotonic, with every update to the document.
If UpsertCAS is called with CAS==0, it implies a CREATE OP.

Please note that UpsertCAS will cost an addition GET op to check for the
current CAS value of the document.