api: can mark a node as deleted.
callbacks on IndexReader{} API should check whether the node is marked
as deleted, using IsDeleted() method.

Min() and Max() APIs will always issue a call back with un-deleted node if
any.

Callers issuing DeleteMin(), DeleteMax(), Delete() methods, it is the
responsibility of the caller to mark the node as deleted, by calling
SetDeadseqno() method on the node.