mvcc: fix chain of snapshots.
For every snapshot-tick a new snapshot is created and inserted
at the head of the snapshot chain.

All write operations happen on the new snapshot head, and remaining
snapshots are considered immutable.

BeginTxn and ViewTxn should obtain the latest immutable snapshot,
which is the snapshot immediately linked with the head snapshot.

Get() can obtain the latest write snapshot, which the head of
the snapshot.

This implies that mvcc.snapshot and snapshot.next should be protected
for concurrent access.

To acquire a new snapshot, don't use locks, instead use a dual-latch
technique using snapshot.refcount and snapshot.purgetry.