Log structured merge (LSM deletes) and ACID compliance.
ACID compliance require that transaction can operate a stable
snapshot that doesn't mutate while the transaction is active.

In case of MVCC, we have an optimization where we don't do copy
on write for Deletes in LSM mode. This would mean it is difficult
to provide stable snapshot for MVCC.

To circumvent this situation we are adding another field called
born-seqno. This will get updated for every insert/update operation.
And dead-seqno will be updated for every delete operation.