llrb: lsm delete for ACID compliance.
LSM deletes where using atomic operation to mark an entry as deleted.
This is very fast, as fast as the GET operation. But this violates
the ACID requirement that transactions can happen only on a stable
snapshot.

This commit adds a method lsm-delete that uses MVCC technique to
statisfy ACID requirements.