bubt: bug fixes and test cases.
testcases TestSnapshotScan1, TestSnapshotScan2, TestSnapshotScan3,
TestCursorYNext1, TestCursorYNext2 where used to catch these bugs.

Bug1:

bubt.Scan() path was failing when Bubt was built with a single
path. Instead of using nextblock() pick up the first block under
first shard.

Bug2:

OpenCursor(), GetNext() and YNext() might leave cur.index out
of bound, that is, it might be -1 or (no-of-entries+1).
So check for bounded index before trying to fetch the entry at
index's offset.

Bug3:

nextblock() was assuming that the last block of the z-file will
end at (filesize-zblocksize), which is wrong. It is at
(filesize-MarkerBlocksize)

Bug4:

zblock.findkey() was returning wrong result (-1, nil, 0, false, false).
when list of entries are like "1173", "1174", "1175", ..., "1200" ..
and findkey is called with "1173000000".

Fixed to return (adjust + 1, nil, 0, false, false).