bogn: bug-fixes.
Bugfix-1:

Bogn.Close() should wait for routines to exit and then
purgesnapshot (to purge memory/disk indexes) before
returning.

Bugfix-2:

If dgm is false, that is full data set is in memory, windupiterator()
should not use latest disk snapshot for final persistence. This is
more of an optimization than bug.