llrb: bug fixes to lockless snapshoting and more.
Bugfix-1:

Iterators returned by Scan() API should be closed (finalized) with
a call like: iter(true /*finish*/).
Fix all test cases to finalize iterators returned by Scan() API.

Bugfix-2:

There can be a race condition which can lead to releasing snapshot
objects back to pool, which it is being referenced by readers. To
mitigate this add a latch mechanism in addition to lock mechanism.
The latch is implemented via, acquiresnapshot() and releasesnapshot()
methods.

acquiresnapshot(), will lock the head of the snapshot so that no
other routines can acquire it, that is, while head is latched to one
routine other routines cannot enter the snapshot chain. The latch
is released after:
a. a new head snapshot is prepended to the chain.
b. write snapshot is acquired by refcounting it.
c. read snapshot is acquired by refcounting it.

refecounting will prevent releasing the snapshot object back to the
pool.